Fix(EraseCluster): Fix erase cluster penValue implementation
by correctly storing and serializing negative 0 (-0.0 can be distinguished from 0.0 in python)

+++ src/ipyniivue/widget.py
@@ -1280,12 +1280,12 @@ class NiiVue(anywidget.AnyWidget):
         self.opts.is_nearest_interpolation = is_nearest
         self.send({"type": "set_interpolation", "data": [is_nearest]})
 
-    def set_pen_value(self, pen_value: int, is_filled_pen: bool):
+    def set_pen_value(self, pen_value: float, is_filled_pen: bool):
         """Determine color and style of drawing.
 
         Parameters
         ----------
-        pen_value : int
+        pen_value : float
             sets the color of the pen
         is_filled_pen : bool
             determines if dragging creates flood-filled shape
@@ -1294,7 +1294,7 @@ class NiiVue(anywidget.AnyWidget):
         --------
         ::
 
-            nv.set_pen_value(1, True)
+            nv.set_pen_value(1.0, True)
         """
         self.opts.pen_value = pen_value
         self.opts.is_filled_pen = is_filled_pen
